提交测试代码
git-svn-id: http://svn01/repositories/r01/ppbox/branches/1.2.0@3442 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ base/SourceBase.h
@@ -51,11 +51,11 @@ namespace ppbox
             SegmentPositionEx()
                 : total_state(not_init)
                 , size_beg(0)
-                , size_end((boost::uint64_t)-1)
+                , size_end(boost::uint32_t(-1))
                 , time_beg(0)
-                , time_end((boost::uint64_t)-1)
+                , time_end(boost::uint32_t(-1))
                 , shard_beg(0)
-                , shard_end((boost::uint64_t)-1)
+                , shard_end(boost::uint32_t(-1))
             {
             }
